% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Descriptive statistics function}
\usage{
descriptives(x, type = "sample", dig = 3, plot = F)
}
\arguments{
\item{x}{Specifies the variable/vector to be analyzed (required).}

\item{type}{Specifies whether the data represents a sample or a population. Default is sample. Abbreviations s and p are ok.}

\item{plot}{Boolean to generate a histogram of the data. Default is FALSE.}
}
\description{
This function calculates a suite of robust descriptive statistics for interval or ratio-scale data.
}
\examples{
x<-runif(100,90,110)
descriptives(x)	#returns an analysis of a distribution with (most likely) no outliers.
x.2<-c(rnorm(100,100,15), 19, 12, 190)
descriptives(x.2, type="p", plot=T)	#returns an analysis of a distribution with outliers using population formulas for variance and std. deviation and asks for the histogram.
}

