% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{Standard Error Summary function}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95)
}
\arguments{
\item{data}{Specifies the dataframe/matrix. Required.}

\item{measurevar}{Specifies the measured variable to be summarized. Required.}

\item{groupvars}{Specifies the grouping variable(s). Required.}

\item{na.rm}{What to do with missing values. Default = FALSE.}

\item{conf.interval}{Specifies the desired confidence interval. Default = 95%}
}
\description{
This function calculates standard errors for groups and summarizes the data for clean plotting with ggplot2.
}
\examples{
x <- data.frame(condition = as.factor(c(rep("control", 50), rep("experimental", 50))), IQ = runif(100,90,110))
summarySE(data = x, measurevar = "IQ", groupvars = "condition")
}

